promiseEach for Promise on iterations.
transform a function which takes a iteration callback and a complete
callback to a function that takes a special iteration callback and
returns a Promise that will be complete when all iterations has been
realized.
This is the same a using a .all()-like callback API converted to a
promise by 'promiseme' but this will prevent from loading a full list in
memory in one time.