New objects model and inheritence.
One should use the Class function to create an object/class now.
It handles inheritence the easy way.
You can pass an object to Class (constructed with { ... } syntax), or a
function that returns an object. This function takes the base
constructor and prototype as arguments than you can use as a closure.
The object constructor is defined using the special __init__ function.

This commit contains also the code for saveOrUpdate for Movie.